pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flex-query"
dynamic = ["version"] 
description = "A flexible, parameter-driven SQL query execution tool that connects to multiple database environments."
readme = "README.md"
authors = [
    {name = "mirosalwsteblik"}
]
requires-python = ">=3.9"
classifiers = [
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pandas",
    "click",
    "pyyaml",
    "sqlalchemy",
    "duckdb-engine",
    "pyodbc",
    "openpyxl",
    "xlsxwriter",
    "markdown"
]

[project.optional-dependencies]
dev = [
    "build",
    "wheel",
    "twine",
    "pytest",
    "black",
    "flake8",
]

[project.scripts]
flexquery = "flexquery.cli.main:cli"

[tool.setuptools]
package-dir = {"" = "core"}

[tool.setuptools.packages.find]
where = ["core"]  # look for packages in the core directory
include = ["flexquery*"]  # include all packages starting with flexquery
exclude = ["tests*", "future*"]  

[tool.setuptools.data-files]
"scripts" = ["core/scripts/*.bat"]

[tool.setuptools.package-data]
"*" = ["*.yml", "*.yaml", "*.sql", "README.md", "docs/*.md", "*.html"]

[tool.setuptools.dynamic]
version = {attr = "flexquery.version.__version__"}

